#2893 Rework a dspace-api IT to use Builders as a demo.
